+++ Sources/StreamChat/Workers/UserListUpdater.swift
@@ -124,6 +124,7 @@ private extension UserPayload {
             isBanned: isBanned,
             isFlaggedByCurrentUser: false,
             userRole: role,
+            teamsRole: teamsRole,
             createdAt: createdAt,
             updatedAt: updatedAt,
             deactivatedAt: deactivatedAt,