+++ Tests/StreamChatUITests/SnapshotTests/ChatChannelList/Search/ChatMessageSearchVC_Tests.swift
@@ -9,6 +9,15 @@ import StreamSwiftTestHelpers
 import XCTest
 
 @MainActor final class ChatMessageSearchVC_Tests: XCTestCase {
+    /// Static setUp() is only run once. Which is what we want in this case to preload the images.
+    override class func setUp() {
+        /// Dummy snapshot to preload the TestImages.yoda.url image
+        /// This was the only workaround to make sure the image always appears in the snapshots.
+        let view = UIImageView(frame: .init(center: .zero, size: .init(width: 100, height: 100)))
+        Components.default.imageLoader.loadImage(into: view, from: TestImages.yoda.url)
+        AssertSnapshot(view, variants: [.defaultLight])
+    }
+
     var mockedClient: ChatClient_Mock!
     var vc: ChatMessageSearchVC!
     var mockedMessageSearchController: ChatMessageSearchController_Mock!
@@ -62,17 +71,17 @@ import XCTest
             .mock(
                 cid: channelId,
                 text: "Some message 1",
-                author: .mock(id: .unique, name: "Yoda")
+                author: .mock(id: .unique, name: "Yoda", imageURL: TestImages.yoda.url)
             ),
             .mock(
                 cid: channelId,
                 text: "Some message 2",
-                author: .mock(id: .unique, name: "Vader")
+                author: .mock(id: .unique, name: "Vader", imageURL: TestImages.yoda.url)
             ),
             .mock(
                 cid: channelId,
                 text: "Some message 3",
-                author: .mock(id: .unique, name: "R2")
+                author: .mock(id: .unique, name: "R2", imageURL: TestImages.yoda.url)
             )
         ]